Refactor turrets to component logic.
Turrets built on uneven floor are now supported server-side. However, the
client-side visualization is not yet correct. The reason is that the precise
orientation of buildable models is computed at the client.

+++ src/sgame/sg_struct.h
@@ -359,7 +359,7 @@ struct gentity_s
 	int         turretLastValidTargetTime;
 	int         turretLastTargetSearch;
 	int         turretLastHeadMove;
-	int         turretCurrentDamage;
+	float       turretCurrentDamage;
 	vec3_t      turretDirToTarget;
 	vec3_t      turretBaseDir;
 	bool    turretDisabled;
